Changed the delete user option to in edit instead of outside.
